Adding my raw log data to Git.
Although adding data to Git is not ideal, I am struggling with where to store this data.

It's awkward to put it on the U drive.

And it's not going to be a lot of data, even after years, so I decided to just put it all in Git for simplicity and ease of continual access.