Updated log through week starting 10.06.2025.
Also added a png of my signature, added the signature to the e-mail, and separated the static version of the email to a separate file called "Send Email."  This separate file includes the actual "send" instructions.

Prep was modified with an error check for too many pipes per line.

Added a "to do" folder and starting to keep files there as well.