Add three double-checks to the notepad log prep script:
Date discontinuity (ensures a date on a new row is never a previous date or two dates in the future)

Another chronological check (by column)

Gaps and overlaps, where the start date on the new row matches the stop date on the old row.  This is calculated per day, and forces me to continuously log all time in a day with no gaps.

Log is committed and email for week starting 10.27.2025 is sent.